requirements.txt
appdirs==1.4.4
asgiref==3.2.10
distlib==0.3.1
Django==3.1
djangorestframework==3.12.2
filelock==3.0.12
importlib-metadata==3.4.0
mysql-connector-python==8.0.17
numpy==1.19.0
ortools==7.7.7810
Pillow==8.1.0
protobuf==3.12.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
virtualenv==20.4.2
zipp==3.4.0
